api/requirements.txt
pydantic
websockets
python-dotenv
azure-identity
fastapi[standard]
azure-ai-inference
opentelemetry-instrumentation
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-fastapi
openai[realtime]
azure-cosmos
aiohttp
aiofiles
types-aiofiles
mypy
azure-ai-projects==1.0.0b10
azure-identity
prompty[azure]==1.0.0a1